게시물 수정, 삭제에 대한 권한체크에 소유권 체크 추가

+++ src/main/java/com/example/sbs/cuni/controller/ArticleController.java
@@ -4,10 +4,8 @@ import java.util.List;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpRequest;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -44,6 +42,17 @@ public class ArticleController {
 
 	@RequestMapping("article/modify")
 	public String showModify(Model model, int id, HttpServletRequest request) {
+		int loginedMemberId = (int) request.getAttribute("loginedMemberId");
+
+		Map<String, Object> articleModifyAvailableRs = articleService.getArticleModifyAvailable(id, loginedMemberId);
+
+		if (((String) articleModifyAvailableRs.get("resultCode")).startsWith("F-")) {
+			model.addAttribute("alertMsg", articleModifyAvailableRs.get("msg"));
+			model.addAttribute("historyBack", true);
+
+			return "common/redirect";
+		}
+
 		Article article = articleService.getArticle(id);
 
 		model.addAttribute("article", article);
@@ -52,10 +61,20 @@ public class ArticleController {
 	}
 
 	@RequestMapping("article/doModify")
-	public String doModify(Model model, @RequestParam Map<String, Object> param) {
-		Map<String, Object> rs = articleService.modify(param);
+	public String doModify(Model model, @RequestParam Map<String, Object> param, HttpServletRequest request) {
+		int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 
 		int id = Integer.parseInt((String) param.get("id"));
+		Map<String, Object> articleModifyAvailableRs = articleService.getArticleModifyAvailable(id, loginedMemberId);
+
+		if (((String) articleModifyAvailableRs.get("resultCode")).startsWith("F-")) {
+			model.addAttribute("alertMsg", articleModifyAvailableRs.get("msg"));
+			model.addAttribute("historyBack", true);
+
+			return "common/redirect";
+		}
+
+		Map<String, Object> rs = articleService.modify(param);
 
 		String msg = (String) rs.get("msg");
 		String redirectUrl = "/article/detail?id=" + id;
@@ -67,7 +86,19 @@ public class ArticleController {
 	}
 
 	@RequestMapping("article/doDelete")
-	public String doDelete(Model model, int id) {
+	public String doDelete(Model model, int id, HttpServletRequest request) {
+
+		int loginedMemberId = (int) request.getAttribute("loginedMemberId");
+
+		Map<String, Object> articleDeleteAvailableRs = articleService.getArticleDeleteAvailable(id, loginedMemberId);
+
+		if (((String) articleDeleteAvailableRs.get("resultCode")).startsWith("F-")) {
+			model.addAttribute("alertMsg", articleDeleteAvailableRs.get("msg"));
+			model.addAttribute("historyBack", true);
+
+			return "common/redirect";
+		}
+
 		Map<String, Object> rs = articleService.deleteArticle(id);
 
 		String msg = (String) rs.get("msg");
